Added: new svelte structure
