+++ app/.gitignore
@@ -7,3 +7,4 @@ node_modules
 .env.*
 !.env.example
 .vercel
+.output
