+++ app/README.md
@@ -8,10 +8,10 @@ If you're seeing this, you've probably already done this step. Congrats!
 
 ```bash
 # create a new project in the current directory
-npm init svelte
+npm create svelte@latest
 
 # create a new project in my-app
-npm init svelte my-app
+npm create svelte@latest my-app
 ```
 
 ## Developing
