+++ app/playwright.config.ts
