+++ app/src/app.d.ts
@@ -1,11 +1,12 @@
-/// <reference types="@sveltejs/kit" />
-
 // See https://kit.svelte.dev/docs/types#app
 // for information about these interfaces
 // and what to do when importing types
 declare namespace App {
-	// interface Locals {}
+	interface Locals {
+		userid: string;
+	}
+
+	// interface PageData {}
+
 	// interface Platform {}
-	// interface Session {}
-	// interface Stuff {}
 }
